usage: correct misleading usage description
When using the -T option the user has to supply two authentication
alogorithms. One for 2G and one for 3G. The authentication algorithms
used to be symbolized by integer numbeers. This was later changed so
that the user could use the clear names of the algorithms. Unfortunately
the usage text was not changed.

- Update usage description

Change-Id: I6a47e4fc24f4b3f0664050032cc15d81703f30c5