sysmo_usimsjs1: Also recognize algo by its number
The test 04_algo.sh suggests that an algorithm should also be
identified by its number (this is the old way). Since we upgraded
the commandline interface to accept the names of the algorithms the
support for the old way that used numbers had been removed. However,
we should stay compatible with the old method to keep legacy scripts
of customers working.

- Check if the given parameter that references tha algorithem is
  a number. If yes, us it, otherwise lookup the number by the
  given string name.

Change-Id: Ib41e4870d05820b6967f648eeeba416113bbc120
Related: OS#3376