utils: allow to use a safe default with str_to_id
The input to str_to_id may originate from user input. It may be that the
input is not found in the given table. In this case an exception is
raised. This may be impractical in some cases. Let's add an optional
safe default that is returned in case the string is not found, this can
be used to either select default settings or to detect an error.

Related: SYS#6473